[GGPUNE-83] Optional inline error message placement on dropdown.
By providing _errorsAboveInput argument, the inline error message will be shown above the select input element, to aid consistency.

If the parameter is not sent, the behaviour will stay the same, with the inline error message below the input.

the error class on the label has also been fixed, as " '  ' + error " was literally rendered. now just " error" is rendered.